requirements.txt
streamlit
gspread
pandas
python-dotenv
requests
matplotlib
